code/lidar/desktop/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=he774522_0
ca-certificates=2019.11.28=hecc5488_0
certifi=2022.12.7=py310haa95532_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
ecal5=5.11.4=pypi_0
fonttools=4.39.0=pypi_0
kiwisolver=1.4.4=pypi_0
libffi=3.4.2=hd77b12b_6
libprotobuf=3.11.4=h1a1b453_0
matplotlib=3.7.1=pypi_0
numpy=1.24.2=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py310haa95532_0
protobuf=3.20.0=pypi_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyserial=3.5=pypi_0
python=3.10.9=h966fe2a_1
python-dateutil=2.8.2=pypi_0
setuptools=65.6.3=py310haa95532_0
six=1.16.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
tzdata=2022g=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=h8cc25b3_0
